Rename lib/utils => cn + move into designSystem
