+++ packages/ui/spa/components/designSystem/accordion.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as AccordionPrimitive from "@radix-ui/react-accordion";
 import { ChevronDown } from "lucide-react";
 
-import { cn } from "../../lib/utils";
+import { cn } from "./cn";
 
 const Accordion = AccordionPrimitive.Root;
