+++ packages/ui/spa/components/designSystem/calendar.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { ChevronDown, ChevronLeft, ChevronRight } from "lucide-react";
 import { DayPicker } from "react-day-picker";
 
-import { cn } from "../../lib/utils";
+import { cn } from "./cn";
 import { Button, buttonVariants } from "./button";
 import * as Select from "@radix-ui/react-select";
