+++ packages/ui/spa/components/designSystem/card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "../../lib/utils";
+import { cn } from "./cn";
 
 const Card = React.forwardRef<
   HTMLDivElement,
