+++ packages/ui/spa/components/designSystem/checkbox.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as CheckboxPrimitive from "@radix-ui/react-checkbox";
 import { Check, Minus } from "lucide-react";
 
-import { cn } from "../../lib/utils";
+import { cn } from "./cn";
 
 const Checkbox = React.forwardRef<
   React.ElementRef<typeof CheckboxPrimitive.Root>,
