+++ packages/ui/spa/components/designSystem/command.tsx
@@ -3,7 +3,7 @@ import { DialogProps } from "@radix-ui/react-dialog";
 import { Command as CommandPrimitive } from "cmdk";
 import { Search } from "lucide-react";
 
-import { cn } from "../../lib/utils";
+import { cn } from "./cn";
 import { Dialog, DialogContent } from "./dialog";
 
 const Command = React.forwardRef<
