+++ packages/ui/spa/components/designSystem/dialog.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
 import { X } from "lucide-react";
 
-import { cn } from "../../lib/utils";
+import { cn } from "./cn";
 
 const Dialog = DialogPrimitive.Root;
