+++ packages/ui/spa/components/designSystem/form.tsx
@@ -10,7 +10,7 @@ import {
   useFormContext,
 } from "react-hook-form";
 
-import { cn } from "../../lib/utils";
+import { cn } from "./cn";
 import { Label } from "./label";
 
 const Form = FormProvider;
