+++ packages/ui/spa/components/designSystem/select.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as SelectPrimitive from "@radix-ui/react-select";
 import { Check, ChevronDown } from "lucide-react";
 
-import { cn } from "../../lib/utils";
+import { cn } from "./cn";
 
 const Select = SelectPrimitive.Root;
